style: change style, fonts, sidebar toggle position. Add username in header
