+++ magner/lib/assets/styles/fonts.css
@@ -1,0 +1,29 @@
+@font-face {
+    font-family: 'SF UI Display';
+    src: local('San Francisco(SF UI) Display Regular'), local('San-Francisco(SF-UI)-Display-Regular'),
+        url('../fonts/SFUIDisplay/SFUIDisplay-Regular.woff2') format('woff2'),
+        url('../fonts/SFUIDisplay/SFUIDisplay-Regular.woff') format('woff'),
+        url('../fonts/SFUIDisplay/SFUIDisplay-Regular.ttf') format('truetype');
+    font-weight: 400;
+    font-style: normal;
+  }
+
+@font-face {
+  font-family: 'SF UI Display';
+  src: local('San Francisco(SF UI) Display Medium'), local('San-Francisco(SF-UI)-Display-Medium'),
+  url('../fonts/SFUIDisplay/SFUIDisplay-Medium.woff2') format('woff2'),
+  url('../fonts/SFUIDisplay/SFUIDisplay-Medium.woff') format('woff'),
+  url('../fonts/SFUIDisplay/SFUIDisplay-Medium.ttf') format('truetype');
+  font-weight: 500;
+  font-style: normal;
+}
+
+@font-face {
+  font-family: 'SF UI Display';
+  src: local('San Francisco(SF UI) Display Light'), local('San-Francisco(SF-UI)-Display-Light'),
+  url('../fonts/SFUIDisplay/SFUIDisplay-Light.woff2') format('woff2'),
+  url('../fonts/SFUIDisplay/SFUIDisplay-Light.woff') format('woff'),
+  url('../fonts/SFUIDisplay/SFUIDisplay-Light.ttf') format('truetype');
+  font-weight: 300;
+  font-style: normal;
+}
