+++ magner/lib/assets/styles/variables.css
@@ -4,7 +4,7 @@
 */
 
 :root {
-  --font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Helvetica Neue, Helvetica, PingFang SC, Hiragino Sans GB, Microsoft YaHei, Arial, sans-serif;
+  --font-family: SF UI Display, -apple-system, BlinkMacSystemFont, Segoe UI, Helvetica Neue, Helvetica, PingFang SC, Hiragino Sans GB, Microsoft YaHei, Arial, sans-serif;
   --font-regular: normal normal 1rem/1.125rem var(--font-family);
   --font-regular-light: 300 normal 1rem/1.125rem var(--font-family);
   --font-h1: 500 normal 1.8rem/2.8rem var(--font-family);
@@ -14,4 +14,6 @@
 
   --color-background: #FFFFFF;
   --color-font-main: #000000;
+  --letter-spacing: 0.0128571em;
+
 }
